+++ src/mappings/prediction-markets/decode.ts
